committed takashis fixes for 64bit and gcc 3.3.  Fixed one minor popup menu leak

+++ src/FTactiveBarGraph.cpp
@@ -566,7 +566,7 @@ void FTactiveBarGraph::xToBinRange(int x, int &frombin, int &tobin)
 			float xscale = x / (float)_width;
 			
 			// use log scale for freq
-			lbin = rbin = (int) pow ( totbins/3.0, xscale) - 1;
+			lbin = rbin = (int) pow ( (float)(totbins/3.0), xscale) - 1;
 			
 			// find lowest with same x
 			while ( (lbin > 0) && ((int)(((FTutils::fast_log10(lbin) / FTutils::fast_log10(totbins/3.0)) * _width)) == x)) {
